requirements.txt
# Core
ultralytics>=8.0.0        # YOLOv8 (includes ByteTrack)
opencv-python>=4.8.0
yt-dlp>=2024.1.0
numpy>=1.24.0
PyYAML>=6.0

# Optional enhancements
matplotlib>=3.7.0
scipy>=1.11.0
pandas>=2.0.0

# Notebook
jupyter>=1.0.0
ipykernel>=6.0.0